scrapers/requirements.txt
python-slugify
python-twitter>=3.0rc1
pyyaml
scrapy>=1.1.0rc
